Add TsLastN functionality to YB Jedis driver.
Summary:
We plan to support TSLastN for our redis time series data type since retrieving the latest
N values in a time series is a pretty common operation. For this purpose, we need to extend the YB
fork of the Jedis driver to include a tsLastN method.

Test Plan: Java tests in YB

Reviewers: hector, robert, rahuldesirazu

Reviewed By: rahuldesirazu

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D4506